refactor 2015; add more tests

+++ adventofcode/src/adventofcode/year2015/Day1.scala
@@ -4,22 +4,22 @@ import adventofcode.Day
 import cats.effect._
 
 object Day1 extends IOApp {
-  case class Context(process: List[Long] => Option[Long])
+  case class Context(process: List[Int] => Option[Int])
 
-  private def partOneSymbolToScore(c: Char): Long = c match {
+  private def partOneSymbolToScore(c: Char): Int = c match {
     case '(' => 1
     case ')' => -1
     case _   => 0
   }
 
-  private def processPartOne(input: List[Long]): Option[Long] =
-    Some(input.foldLeft(0L)(_ + _))
+  private def processPartOne(input: List[Int]): Option[Int] =
+    Some(input.foldLeft(0)(_ + _))
 
-  private def processPartTwo(input: List[Long]): Option[Long] =
-    Some(input.foldLeft((0L, 0L))((a, b) => if (a._2 < 0) (a._1, a._2) else (a._1 + 1, a._2 + b))._1)
+  private def processPartTwo(input: List[Int]): Option[Int] =
+    Some(input.foldLeft((0, 0))((a, b) => if (a._2 < 0) (a._1, a._2) else (a._1 + 1, a._2 + b))._1)
 
-  object Runner extends Day[List[Long], Context, Long](2015, 1) {
-    override def transformInput(lines: List[String]): List[Long] =
+  object Runner extends Day[List[Int], Context, Int](2015, 1) {
+    override def transformInput(lines: List[String]): List[Int] =
       lines.mkString.trim.map(partOneSymbolToScore).toList
 
     override def partOneContext(): Option[Context] =
@@ -28,7 +28,7 @@ object Day1 extends IOApp {
     override def partTwoContext(): Option[Context] =
       Some(Context(processPartTwo))
 
-    override def process(input: List[Long], context: Option[Context]): Option[Long] =
+    override def process(input: List[Int], context: Option[Context]): Option[Int] =
       context.flatMap(_.process(input))
   }
