+++ adventofcode/src/adventofcode/year2015/Day6.scala
@@ -31,9 +31,9 @@ object Day6 extends IOApp {
     }
   }
 
-  case class Context(turnOn: Long => Long, turnOff: Long => Long, toggle: Long => Long)
+  case class Context(turnOn: Int => Int, turnOff: Int => Int, toggle: Int => Int)
 
-  object Runner extends Day[List[SantaInstruction], Context, Long](2015, 6) {
+  object Runner extends Day[List[SantaInstruction], Context, Int](2015, 6) {
     override def transformInput(lines: List[String]): List[SantaInstruction] =
       lines.map(SantaInstruction.from)
 
@@ -43,9 +43,9 @@ object Day6 extends IOApp {
     override def partTwoContext(): Option[Context] =
       Some(Context(i => i + 1, i => if ((i - 1) < 0) 0 else i - 1, i => i + 2))
 
-    override def process(input: List[SantaInstruction], context: Option[Context]): Option[Long] =
+    override def process(input: List[SantaInstruction], context: Option[Context]): Option[Int] =
       context.map { ctx =>
-        val grid = Array.ofDim[Long](1000, 1000)
+        val grid = Array.ofDim[Int](1000, 1000)
         input.foreach { instruction =>
           for (x <- (instruction.range.x1 to instruction.range.x2))
             for (y <- (instruction.range.y1 to instruction.range.y2))
